Added pivot table usersToCourse. Moved copies of all tables and queries into users.js for testing.
