--- models/cource.js
@@ -1,24 +1,0 @@
-// Import the ORM to create functions that will interact with the database.
-var orm = require("../config/orm.js");
-
-var song = {
-  all: function(cb) {
-    orm.all("songs", function(res) {
-      cb(res);
-    });
-  },
-  // The variables cols and vals are arrays.
-  create: function(cols, vals, cb) {
-    orm.create("songs", cols, vals, function(res) {
-      cb(res);
-    });
-  },
-  update: function(objColVals, condition, cb) {
-    orm.update("songs", objColVals, condition, function(res) {
-      cb(res);
-    });
-  }
-};
-
-// Export the database functions for the controller (catsController.js).
-module.exports = song;
